+++ src/app/modules/step-0-settings-page/initialize-spinner/initialize-spinner.component.ts
@@ -1,0 +1,11 @@
+import { Component } from '@angular/core';
+
+/**
+ * Initialization spinner component.
+ */
+@Component({
+  selector: 'app-initialize-spinner',
+  templateUrl: './initialize-spinner.component.html',
+  styleUrls: ['./initialize-spinner.component.less']
+})
+export class InitializeSpinnerComponent {}